Create calculate_G.m
Code to calculate G from ROMS output. The shear rate, G, to estimate floc shear aggregation and erosion, is calculated as G=sqrt(diss/nu), where nu is 1.5 10-6 and diss is the dissipation. The dissipation in the water column (except in the bottommost layer) is estimated from the GLS mixing (k-omega). The dissipation in the bottommost layer is enhanced to take the effect of the wave bottom boundary layer. In that layer, the dissipation is calculated as diss = ustr^(1.5)/(vonKar*0.9), where ustr is the magnitude of the wave-current bottom stress. This formulation is now part of the FLOC_BBL_DISS cpp option.